+++ DevOps/States/BacklogState/IBacklogState.cs
@@ -1,4 +1,4 @@
-﻿namespace DevOps.States.Backlog {
+﻿namespace DevOps.States.BacklogState {
     public interface IBacklogState {
 
         public void StartTask();
