+++ DevOps/States/BacklogState/TestedState.cs
@@ -1,15 +1,11 @@
 ﻿using DevOps.Domain;
 
-namespace DevOps.States.Backlog {
-    public class TodoState : IBacklogState {
+namespace DevOps.States.BacklogState {
+    public class TestedState : IBacklogState {
         private readonly BacklogItem _backlogItem;
 
-        public TodoState(BacklogItem backlogItem) {
+        public TestedState(BacklogItem backlogItem) {
             _backlogItem = backlogItem;
-        }
-
-        public void StartTask() {
-            _backlogItem.UpdateState(new DoingState(_backlogItem));
         }
 
         public int FinishTask() => 0;
@@ -19,10 +15,18 @@ namespace DevOps.States.Backlog {
         }
 
         public void ReviewTestReport(bool passed) {
-            Console.WriteLine("Item hasn't started yet..");
+            if (!passed) {
+                _backlogItem.UpdateState(new ReadyForTestingState(_backlogItem));
+            } else {
+                _backlogItem.UpdateState(new DoneState(_backlogItem));
+            }
         }
 
         public int SendTestReport(bool passed) => 0;
+
+        public void StartTask() {
+            Console.WriteLine("Item hasn't started yet..");
+        }
 
         public void StartTesting() {
             Console.WriteLine("Item hasn't started yet..");
