+++ DevOps/States/BacklogState/TestingState.cs
@@ -1,24 +1,32 @@
 ﻿using DevOps.Domain;
 
-namespace DevOps.States.Backlog {
-    public class DoneState : IBacklogState {
+namespace DevOps.States.BacklogState {
+    public class TestingState : IBacklogState {
         private readonly BacklogItem _backlogItem;
 
-        public DoneState(BacklogItem backlogItem) {
+        public TestingState(BacklogItem backlogItem) {
             _backlogItem = backlogItem;
         }
 
         public int FinishTask() => 0;
 
         public void InvalidateTask() {
-            _backlogItem.UpdateState(new TodoState(_backlogItem));
+            Console.WriteLine("Item hasn't started yet..");
         }
 
         public void ReviewTestReport(bool passed) {
             Console.WriteLine("Item hasn't started yet..");
         }
 
-        public int SendTestReport(bool passed) => 0;
+        public int SendTestReport(bool passed) {
+            if (!passed) {
+                _backlogItem.UpdateState(new TodoState(_backlogItem));
+                return _backlogItem.NotifyScrumMaster();
+            } else {
+                _backlogItem.UpdateState(new TestedState(_backlogItem));
+                return 0;
+            }
+        }
 
         public void StartTask() {
             Console.WriteLine("Item hasn't started yet..");
@@ -29,7 +37,7 @@ namespace DevOps.States.Backlog {
         }
 
         public void StopTask() {
-            _backlogItem.UpdateState(new DoneState(_backlogItem));
+            Console.WriteLine("Item hasn't started yet..");
         }
 
         public void StopTesting() {
